chore: FilterTabs and FilterActions
