+++ packages/frontend/src/components/DashboardFilter/FilterConfiguration/types.ts
@@ -1,0 +1,9 @@
+export enum FilterActions {
+    ADD = 'add',
+    REMOVE = 'remove',
+}
+
+export enum FilterTabs {
+    SETTINGS = 'settings',
+    TILES = 'tiles',
+}
